+++ src/lib/BlogPostHeader.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import Feed_Item_Date from './Feed_Item_Date.svelte';
-	import type {Blog_Post_Item} from './blog.js';
+	import FeedItemDate from './FeedItemDate.svelte';
+	import type {BlogPostItem} from './blog.js';
 
 	interface Props {
-		item: Blog_Post_Item;
+		item: BlogPostItem;
 	}
 
 	const {item}: Props = $props();
@@ -12,6 +12,6 @@
 <header>
 	<h1 class="mt_xl5 mb_lg">{item.title}</h1>
 	<p class="mb_xl5">
-		<Feed_Item_Date {item} />
+		<FeedItemDate {item} />
 	</p>
 </header>