+++ src/lib/post.task.ts
@@ -1,4 +1,4 @@
-import {Task_Error, type Task} from '@ryanatkn/gro';
+import {TaskError, type Task} from '@ryanatkn/gro';
 import {z} from 'zod';
 import {format_file} from '@ryanatkn/gro/format_file.js';
 import {mkdir, writeFile} from 'node:fs/promises';
@@ -28,7 +28,7 @@ export const task: Task<Args> = {
 		} = args;
 
 		if (!raw_title) {
-			throw new Task_Error('post title is required, e.g. `gro post "Hello world"`');
+			throw new TaskError('post title is required, e.g. `gro post "Hello world"`');
 		}
 		const title = raw_title.trim();
 		const slug = slugify(title);
@@ -52,7 +52,7 @@ export const task: Task<Args> = {
 
 		const unformatted = `
 			<script lang="ts" module>
-				import type {Blog_Post_Data} from '${fuz_blog_import_path}/blog.js';
+				import type {BlogPostData} from '${fuz_blog_import_path}/blog.js';
 
 				export const post = {
 					title: ${JSON.stringify(title)},
@@ -61,19 +61,19 @@ export const task: Task<Args> = {
 					date_modified: '${date}',
 					summary: 'todo',
 					tags: ['todo'],
-				} satisfies Blog_Post_Data;
+				} satisfies BlogPostData;
 			</script>
 
 			<script lang="ts">
-				import Blog_Post from '${fuz_blog_import_path}/Blog_Post.svelte';
+				import BlogPost from '${fuz_blog_import_path}/BlogPost.svelte';
 			</script>
 
 			<!-- This component is totally optional, you have full control over the page. -->
-			<Blog_Post {post}>
+			<BlogPost {post}>
 				<p>
 					TODO content goes here
 				</p>
-			</Blog_Post>
+			</BlogPost>
 		`;
 		const formatted = await format_file(unformatted, {parser: 'svelte'});