+++ src/routes/blog/+layout.svelte
@@ -5,7 +5,7 @@
 	import {DEV} from 'esm-env';
 	import type {Snippet} from 'svelte';
 	import {
-		Mastodon_Cache,
+		MastodonCache,
 		mastodon_cache_context,
 	} from '@ryanatkn/fuz_mastodon/mastodon_cache.svelte.js';
 
@@ -17,7 +17,7 @@
 
 	if (DEV) {
 		mastodon_cache_context.set(
-			new Mastodon_Cache(
+			new MastodonCache(
 				async () => (await import('./mastodon_dev_cache_data.js')).mastodon_dev_cache_data,
 			),
 		);
@@ -38,7 +38,7 @@
 		// as the exported `mastodon_dev_cache_data` value.
 	}
 
-	// TODO redirect from the numbers or render UI to navigate to it, maybe in `Blog_Post`
+	// TODO redirect from the numbers or render UI to navigate to it, maybe in `BlogPost`
 </script>
 
 <div class="breadcrumbs">