+++ src/routes/docs/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Docs_Content from '@ryanatkn/fuz/Docs_Content.svelte';
+	import DocsContent from '@ryanatkn/fuz/DocsContent.svelte';
 	import {tomes_context} from '@ryanatkn/fuz/tome.js';
 	import {pkg_context} from '@ryanatkn/fuz/pkg.svelte.js';
 
@@ -10,4 +10,4 @@
 	const pkg = pkg_context.get();
 </script>
 
-<Docs_Content {tomes} {pkg}></Docs_Content>
+<DocsContent {tomes} {pkg}></DocsContent>